Fix webdriver-based scraping (closes #8). Adds logs. Adds testing mode.
Key to fixing webdriver-based scraping seems to be implementing a brief wait before checking if the file exists. Also added the ability for the script to retry, but this may not be necessary and could be removed in the future.

Success and failure to download particular files are now noted in the log. Failure to download a file no longer results in the entire script stopping.

Testing mode (server-side with argument test or on a local machine with argument localtest) allows downloads to be verified without being committed to GitHub (in fact, the PyGithub module is ignored completely in testing mode).