Add "official" versions of MB datasets. #120
Also rename some datasets to align with their "official" names.

This may also solve the record limit issue for some datasets. #196

However, it is unclear if these datasets will always be the most-to-date as some CSVs require manual regeneration. For these, the FGDB versions seem to be always the newest version.