Retire BC public exposure flight notifications.
From the website:

As of November 16, 2021, exposure notifications for passenger travel by plane, train, bus or cruise ship will no longer be posted to the BCCDC website.

All passengers are required to be fully vaccinated to board a plane, train or cruise ship for domestic travel in Canada as well as to enter Canada from international travel. Anyone who is travelling should monitor their health before, during and after travel. If you develop symptoms, self-isolate and get tested.

Archived flight exposure notifications will be available on this page until the end of November 2021.